División en Tablero Nuevos archivos para la modularización del tablero.
