+++ func/config/idioma.js
@@ -1,5 +1,5 @@
 import { traducir_editor } from "../modo/editor/editor.js";
-import { traducir_tablero } from "../modo/tablero.js";
+import { traducir_tablero } from "../modo/tablero/tablero.js";
 import { traducir_visor } from "../visor/visor.js";
 import { cambiar_cabecera as traducir_cabecera } from "../util/cabecera.js";
 import carga_json from "../util/json.js";
