+++ index.html
@@ -35,13 +35,7 @@
 	<main>
 		<div style="height: 20px;"></div>
 		<div id="panel">
-			<div id="botones" class="botones">
-				<button id="cargar_tempo" class="botón"> Cargar Tempo </button>
-				<input id="cargar_real" class="oculto" type="file" accept=".json" multiple>
-				<button id="descargar_todo" class="botón"> Descargar Todo </button>
-				<button id="borrar_todo" class="botón"> Borrar Todo </button>
-			</div>
-
+			<div id="botones" class="botones"></div>
 			<div id="tablero"></div>
 		</div>
 	</main>
